// client-id/src/app/core/models/test-share-models/process-temp-save.ts
export class ProcessTempSave{
    inspecTemp: InspecTemp;
    inspecCnfrmnIssu: InspecCnfrmnIssu;
    userId: string;
    constructor(){
        this.inspecTemp = new InspecTemp();
        this.inspecCnfrmnIssu = new InspecCnfrmnIssu();
        this.userId = null;
    }
  }
// ...
export class InspecCnfrmnIssu{
    athrzSn: number;
    constructor(){
        this.athrzSn = 0;
    }
}